requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pyPDF2
faiss-cpu
langchain_google_genai 
flask